+++ src/main/java/org/example/pdf_creator/content/MainSection.java
@@ -14,6 +14,10 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 import org.example.pdf_creator.content.abstractsclasses.TextSectionList;
 
+/**
+ * Type of TextSection which can have its own icon, create a line divider between
+ * this section and another one below it, and has its own TextSectionList
+ */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor