+++ src/main/java/org/example/pdf_creator/content/PdfCreationConfiguration.java
@@ -1,18 +1,28 @@
 package org.example.pdf_creator.content;
 
 import com.itextpdf.layout.element.BlockElement;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.example.pdf_creator.content.abstractsclasses.TextSection;
 import org.example.pdf_creator.content.abstractsclasses.TextSectionList;
 import org.example.pdf_creator.factories.TextSectionListType;
 
 import java.util.List;
 
+/**
+ * Basic representation of TextSectionList.
+ * Mostly used for listing of top-most sections.
+ */
 @Data
 public class PdfCreationConfiguration extends TextSectionList {
 
     public PdfCreationConfiguration(List<TextSection> textSectionList, List<FontPreset> fontPresets) {
         super(textSectionList, fontPresets);
+    }
+
+    public PdfCreationConfiguration() {
+        super();
     }
 
     @Override